+++ src/zen_do/__init__.py
@@ -2,13 +2,11 @@
 
 from .cli import zenodo_publish
 from .examples import example_deposit, example_metadata
-from .zenodo import (
+from .zenodo_client import (
     ZenodoClient,
-    ZenodoCreator,
     ZenodoDepositState,
-    ZenodoMetadata,
-    ZenodoRelatedIdentifier,
 )
+from .zenodo_metadata import ZenodoCreator, ZenodoMetadata, ZenodoRelatedIdentifier
 
 __all__ = [
     "example_metadata",
