+++ src/zen_do/zenodo_client.py
@@ -1,72 +1,10 @@
-import re
 from enum import StrEnum
 from pathlib import Path
-from typing import Any, Optional, Self, Union
+from typing import Any, Union
 
 import requests
-import seedcase_soil as so
-from pydantic import BaseModel, model_validator
 
-
-class ZenodoCreator(BaseModel):
-    """Model representing the creator of a Zenodo deposit.
-
-    Attributes:
-        name: The name of the creator.
-        affiliation: The affiliation of the creator.
-        orcid: The ORCID of the creator.
-    """
-
-    name: str
-    affiliation: str
-    orcid: str
-
-
-class ZenodoRelatedIdentifier(BaseModel):
-    """Model representing an identifier related to a Zenodo deposit.
-
-    Attributes:
-        identifier: The value of the identifier.
-        relation: The relationship between the deposit and the other piece of work
-            identified by the identifier.
-        resource_type: The type of the work identified by the identifier.
-        scheme: The scheme followed by the identifier.
-    """
-
-    identifier: str
-    relation: str
-    resource_type: str
-    scheme: Optional[str] = None
-
-    @model_validator(mode="after")
-    def _check_urn(self) -> Self:
-
-        if self.scheme == "urn" and not re.fullmatch(
-            r"urn:zenodo(:[^/:]+)+", self.identifier
-        ):
-            raise ValueError(
-                f"The URN {self.identifier!r} does not have the expected format. URNs "
-                "must be in the format 'urn:zenodo:<unique-id>(:<optional-sub-id>)'. "
-                "We recommend 'urn:zenodo:<github-username>:<repo-name>:<output-type>'."
-            )
-        return self
-
-
-class ZenodoMetadata(BaseModel):
-    """Model representing Zenodo metadata.
-
-    Attributes:
-        title: The title of the deposit.
-        upload_type: The type of the deposit.
-        creators: The creators of the deposit.
-        related_identifiers: Identifiers related to the deposit.
-    """
-
-    title: str
-    upload_type: str
-    creators: list[ZenodoCreator]
-    related_identifiers: list[ZenodoRelatedIdentifier] = []
-
+from zen_do.zenodo_metadata import ZenodoMetadata
 
 type ZenodoResponse = dict[str, Any]
 
@@ -95,66 +33,6 @@ def _is_deposit_editable(deposit: ZenodoResponse) -> bool:
         ZenodoDepositState.inprogress,
         ZenodoDepositState.unsubmitted,
     ]
-
-
-def zenodo_get_deposit(deposits: list[ZenodoResponse]) -> Optional[ZenodoResponse]:
-    """Gets the Zenodo deposit for the repository if it exists.
-
-    Gets the URN identifier from the `.zenodo.json` file. If one
-    doesn't exist, this function will not work.
-
-    Args:
-        deposits: All the deposits on Zenodo associated with an access token.
-
-    Returns:
-        The Zenodo deposit for the repo if it exists, None otherwise.
-    """
-    urn = _get_urn()
-
-    matching_deposits = so.keep(
-        deposits,
-        lambda deposit: bool(
-            so.keep(
-                _get_zenodo_field(deposit, "metadata").get("related_identifiers", []),
-                lambda id: _urn_matches(id, urn),
-            )
-        ),
-    )
-
-    if len(matching_deposits) > 1:
-        raise ValueError(
-            "There are multiple deposits on Zenodo with the URN "
-            f"{urn!r} as a `related identifier`. We only allow one."
-        )
-    if not matching_deposits:
-        return None
-    return matching_deposits[0]
-
-
-def _load_zenodo_json() -> ZenodoMetadata:
-    return ZenodoMetadata.model_validate_json(Path(".zenodo.json").read_text())
-
-
-def _urn_matches(id_response: ZenodoResponse, target_urn: str) -> bool:
-    id = ZenodoRelatedIdentifier.model_construct(**id_response)
-    return _is_urn(id) and id.identifier == target_urn
-
-
-def _is_urn(id: ZenodoRelatedIdentifier) -> bool:
-    return id.relation == "isIdenticalTo" and id.scheme == "urn"
-
-
-def _get_urn() -> str:
-    metadata = _load_zenodo_json()
-    ids = so.keep(metadata.related_identifiers, _is_urn)
-    if len(ids) != 1:
-        raise ValueError(
-            "Expected exactly one `isIdenticalTo` URN in `.zenodo.json` under "
-            f"`related_identifiers`, but found {len(ids)}. Ensure there is a single "
-            "unique URN, as it is used to identify the corresponding deposit on Zenodo."
-        )
-
-    return ids[0].identifier
 
 
 class ZenodoClient:
